requirements.txt
alembic==1.7.5
Authlib==0.15.5
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
pycparser==2.21
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
requests==2.26.0
simple-websocket==0.5.0
six==1.16.0
SQLAlchemy==1.4.28
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
